Separate app logic from dom module
